Hypothesis-Testing
The "Data Science Hypothesis Testing" repository is an extensive and in-depth resource that aims to provide a comprehensive understanding of hypothesis testing in the field of data science. Whether you are a novice or an experienced data scientist, this repository covers a wide range of topics, including fundamental concepts, practical techniques, and real-world examples.

Starting with an introduction to hypothesis testing, the repository delves into the key statistical concepts required for hypothesis testing, such as population vs. sample, sampling distribution, and the Central Limit Theorem. It explores the different types of hypotheses, including null and alternative hypotheses, and discusses the importance of controlling Type I and Type II errors.

The repository explores various one-sample and two-sample hypothesis tests, including z-tests, t-tests, and hypothesis testing for proportions. It also delves into more advanced topics like ANOVA (Analysis of Variance) and non-parametric hypothesis tests.

Additionally, the repository covers multiple comparisons and post-hoc testing techniques, such as the Bonferroni correction and Tukey's HSD test. It discusses the assumptions and robustness of hypothesis tests, as well as the significance of effect size and power analysis in data analysis.

Practical examples and case studies are provided to demonstrate the application of hypothesis testing in different domains, such as A/B testing, medical research, social sciences, and business analytics.

This repository encourages contributions from the community to expand and improve the content. It is licensed under the MIT License and is for educational purposes only.